Refactor: delete `TotalStakeWidget` file
