+++ src/modules/dashboard/components/ActionsPage/TotalStakeWidget/index.ts
@@ -1,1 +1,2 @@
-export { default, StakeSide } from './TotalStakeWidget';
+export { default as SingleTotalStake } from './SingleTotalStake';
+export { default as GroupedTotalStake } from './GroupedTotalStake';
